[refactor] init save
